Move assign_registers_new to assign_registers + fix snippets tests
