+++ src/common/snippets/tests/src/registers.cpp
@@ -33,6 +33,8 @@ TEST(TransformationTests, AssignRegisters) {
         auto s00 = std::make_shared<snippets::isa::Store>(y02); s00->set_friendly_name("y03");
         s00->set_friendly_name("s00");
         f = std::make_shared<Function>(NodeVector{s00}, ParameterVector{p0, p1});
+        // Note that testing the result is not strictly necessary, since the Result doesn't emit any code
+        f->get_result()->set_friendly_name("r00");
 
         pass::Manager m;
         m.register_pass<pass::InitNodeInfo>();
@@ -52,18 +54,19 @@ TEST(TransformationTests, AssignRegisters) {
             {"y01", 1},
             {"y02", 2},
             {"s00", 2}, // gpr
+            {"r00", 2}  // gpr
         };
 
         auto total_ops = 0;
         for (auto& op : f->get_ordered_ops()) {
-            auto& rt = op->get_rt_info();
-
-            auto it_rinfo = rt.find("reginfo");
-            if (it_rinfo != rt.end()) {
-                auto reginfo = it_rinfo->second.as<std::vector<size_t>>();
-                auto reg = reginfo[0];
-                ASSERT_TRUE(ref_registers[op->get_friendly_name()] == reg);
-                total_ops++;
+            for (const auto& output : op->outputs()) {
+                const auto& rt = output.get_tensor_ptr()->get_rt_info();
+                auto it_rt = rt.find("reginfo");
+                if (it_rt != rt.end()) {
+                    auto reg = it_rt->second.as<size_t>();
+                    ASSERT_TRUE(ref_registers[op->get_friendly_name()] == reg);
+                    total_ops++;
+                }
             }
         }
         ASSERT_EQ(total_ops, ref_registers.size());
@@ -120,6 +123,7 @@ TEST(TransformationTests, AssignRegisters2) {
         s00->set_friendly_name("s00");
 
         f = std::make_shared<Function>(NodeVector{s00}, ParameterVector{p0, p1, p2, p3, p4, p5, p6, p7});
+        f->get_result()->set_friendly_name("res00");
 
         pass::Manager m;
         m.register_pass<pass::InitNodeInfo>();
@@ -140,17 +144,19 @@ TEST(TransformationTests, AssignRegisters2) {
             {"r18", 0}, {"r19", 2}, {"r20", 4}, {"r21", 1}, {"r22", 0}, {"r23", 6},
             {"r24", 1},
             {"s00", 8},
+            {"res00", 8}
         };
 
         auto total_ops = 0;
         for (auto& op : f->get_ordered_ops()) {
-            auto& rt = op->get_rt_info();
-            auto it_rinfo = rt.find("reginfo");
-            if (it_rinfo != rt.end()) {
-                auto reginfo = it_rinfo->second.as<std::vector<size_t>>();
-                auto reg = reginfo[0];
-                ASSERT_TRUE(ref_registers[op->get_friendly_name()] == reg);
-                total_ops++;
+            for (const auto& output : op->outputs()) {
+                const auto& rt = output.get_tensor_ptr()->get_rt_info();
+                auto it_rt = rt.find("reginfo");
+                if (it_rt != rt.end()) {
+                    auto reg = it_rt->second.as<size_t>();
+                    ASSERT_TRUE(ref_registers[op->get_friendly_name()] == reg);
+                    total_ops++;
+                }
             }
         }
         ASSERT_EQ(total_ops, ref_registers.size());
